+++ Sources/NIO/SelectableChannel.swift
@@ -49,17 +49,3 @@ internal protocol SelectableChannel: Channel {
 
     func reregister(selector: Selector<NIORegistration>, interested: SelectorEventSet) throws
 }
-
-/// Multicast is not supported on this interface.
-public struct NIOMulticastNotSupportedError: Error {
-    public var device: NIONetworkDevice
-
-    public init(device: NIONetworkDevice) {
-        self.device = device
-    }
-}
-
-/// Multicast has not been properly implemented on this channel.
-public struct NIOMulticastNotImplementedError: Error {
-    public init() {}
-}
